u32 everywhere

+++ db/src/test_storage.rs
@@ -54,11 +54,11 @@ impl TestStorage {
 }
 
 impl Store for TestStorage {
-	fn best_block_number(&self) -> Option<u64> {
-		self.data.read().best_block_number.map(|b| b as u64)
+	fn best_block_number(&self) -> Option<u32> {
+		self.data.read().best_block_number.map(|b| b as u32)
 	}
 
-	fn block_hash(&self, number: u64) -> Option<H256> {
+	fn block_hash(&self, number: u32) -> Option<H256> {
 		let data = self.data.read();
 		data.heights.get(&(number as usize)).cloned()
 	}
